Scaled-Prediction-Models/requirements.txt
# Core Dependencies
pandas>=1.5.0
numpy>=1.23.0
scikit-learn>=1.2.0
xgboost>=1.7.0

# Database
psycopg2-binary>=2.9.0

# Configuration
pyyaml>=6.0
python-dotenv>=0.19.0

# Matrix Imputation
# Optional: Install fancyimpute for Soft-Impute method
# fancyimpute>=0.7.0
# If not installed, will fall back to sklearn's IterativeImputer

# Utilities
python-dateutil>=2.8.0